+++ src/components/contact/contact.css
@@ -1,14 +1,16 @@
 main {
 	position: relative;
-	top: 10px;
+	top: 0;
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
 	align-items: center;
-  padding: 30px 0px;
+	padding: 30px 0px;
+	width: 100%;
+	box-sizing: border-box;
 }
 
-contact-form {
+.contact-form {
 	background: #ffffff;
 	max-width: 800px;
 	height: 100%;
