+++ src/pages/About/about.css
@@ -1,29 +1,85 @@
-.about{
-     /* background-color: #edf3f3; */
-     max-width: 80%;
-     /* border: 2px solid #ccc; */
-     font-weight: 300;
-     margin: 80px 15px;
-     padding: 30px;
+.about {
+	position: relative;
+	height: 100%;
+	width: 100%;
+	/* border: 2px solid #ccc; */
+	font-weight: 300;
+	margin: 0px;
+	box-sizing: border-box;
 }
 
-.about h2{
-color: #193c5a;
-  font-size: 2em;
-  font-weight: 600;
-  text-align: center;
-  margin: 20px 0px;
+.about-container {
+	position: relative;
+	min-height: 70vh;
+	background: url('../../images/home-background.jpg');
+	background-position: bottom left;
+	background-repeat: no-repeat;
+	background-attachment: fixed;
+	background-size: cover;
+	display: flex;
+	justify-content: flex-end;
+	align-items: flex-start;
+	flex-direction: column;
+	color: #ffeeee;
+	z-index: 990;
 }
 
-.about p{
-    font-size: 1.5em;
+.overlay {
+	height: 100%;
+	width: 100%;
+	z-index: 991;
+	background-color: rgba(0, 0, 0, 0.4);
+	position: absolute;
 }
 
-.about-btn{
-    position: relative;
-    left: 40px;
+.about-content {
+	justify-self: flex-end;
+	height: 100%;
+	min-height: 300px;
+	width: 40%;
+	z-index: 992;
+	padding: 20px;
 }
 
-.sub-about{
-    text-align: center;
+.about h2 {
+	font-size: 2em;
+	font-weight: 600;
+	padding: 0px 30px;
+}
+
+.about p {
+	font-size: 12pt;
+	padding: 0px 30px;
+}
+
+.about-btn {
+	position: relative;
+	padding: 0px 30px;
+	margin-bottom: 20px;
+}
+
+.sub-about {
+	text-align: center;
+}
+
+/*------------------------------------------------
+Mobile screen
+---------------------------------------------------*/
+
+@media all and (max-width: 700px) {
+	.about-container {
+		min-height: 100vh;
+		justify-content: center;
+		align-items: flex-start;
+	}
+	.about-content {
+		top: 100px;
+		position: relative;
+		justify-self: flex-start;
+		text-align: center;
+		width: 100%;
+		height: 100%;
+		z-index: 992;
+		padding: 20px;
+	}
 }
