+++ src/pages/landing/App.css
@@ -56,12 +56,12 @@
 	border: 2px solid #f26627;
 }
 .land-intro {
-  margin-top: 5rem;
-	font-family: 'Roboto', serif;
+	margin-top: 5rem;
+	font-family: 'Roboto', cursive;
 	font-size: 45px;
-  font-weight: 600;
+	font-weight: 600;
 	text-align: center;
-	color: #6e98b4;
+	color: #efeeee;
 }
 
 /*------------------------------------------------
